requirements.txt
dj-database-url==0.5.0
Django==2.1.7
django-extensions==2.1.6
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==19.9.0
psycopg2==2.7.7
pytz==2018.9
six==1.12.0
whitenoise==4.1.2
